fix error response and cms logic #2

+++ src/api/services/category.service.js
@@ -20,7 +20,18 @@ const findAll = async (params) => {
 const findOne = async (params) => {
   try {
     const { id } = params;
-
+    if (!id) {
+      throw new CustomAPIError(
+        "Invalid please provide all of the required fields",
+        400
+      );
+    }
+    const existingCategory = await prisma.subCategory.findUnique({
+      where: { id: +id },
+    });
+    if (!existingCategory) {
+      throw new CustomAPIError(`No sub category with id of ${id}`, 400);
+    }
     const categories = prisma.category.findUnique({
       where: {
         id: +id,
@@ -51,7 +62,18 @@ const update = async (pathParams, params) => {
   try {
     const { id } = pathParams;
     const { name } = params;
-
+    if (!id || !name) {
+      throw new CustomAPIError(
+        "Invalid please provide all of the required fields",
+        400
+      );
+    }
+    const existingCategory = await prisma.subCategory.findUnique({
+      where: { id: +id },
+    });
+    if (!existingCategory) {
+      throw new CustomAPIError(`No sub category with id of ${id}`, 400);
+    }
     const categories = prisma.category.update({
       where: {
         id: +id,
@@ -70,7 +92,18 @@ const update = async (pathParams, params) => {
 const destroy = async (params) => {
   try {
     const { id } = params;
-
+    if (!id) {
+      throw new CustomAPIError(
+        "Invalid please provide all of the required fields",
+        400
+      );
+    }
+    const existingCategory = await prisma.subCategory.findUnique({
+      where: { id: +id },
+    });
+    if (!existingCategory) {
+      throw new CustomAPIError(`No category with id of ${id}`, 400);
+    }
     const categories = prisma.category.delete({
       where: {
         id: +id,
